+++ HISTORY.rst
@@ -9,6 +9,14 @@ Changelog
 
 version 1.6.0-dev
 ---------------------------
++ Add a ``--git-aware`` or ``--ga`` option to only copy copy files listed by
+  git ls-files. This omits the ``.git`` folder, all untracked files and
+  everything ignored by ``.gitignore``. This reduces the number of copy
+  operations drastically.
+
+  Pytest-workflow will now emit a warning when copying of a git directory is
+  detected without the ``--git-aware`` option.
+
 + Add support and tests for Python 3.10
 
 version 1.5.0